Clean up test_equal_fields
